# mcp/openapi.yml
openapi: 3.0.3
info:
  title: CivOne MCP HTTP Tool API
  version: 1.1.0
  description: |
    OpenAPI facade for CivOne MCP tools over HTTP.
    Endpoint path can be /mcp or /mcp/<anything>, but the body MUST be a JSON-RPC request.
    Prefilled examples below already contain a valid tools/call envelope.
servers:
  - url: http://127.0.0.1:8765
    description: Local MCP HTTP endpoint
...
paths:
...
  /mcp/game_get_settings:
    post:
      tags: [MCP Tools]
      operationId: game_get_settings
      summary: Read grouped runtime/game settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JsonRpcToolCallRequest'
            example:
              jsonrpc: '2.0'
              id: '1'
              method: tools/call
              params:
                name: game_get_settings
                arguments: {}
      responses:
        '200':
          description: JSON-RPC response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JsonRpcResponse'
...
components:
  schemas:
    JsonRpcToolCallRequest:
      type: object
      required: [jsonrpc, id, method, params]
      properties:
        jsonrpc:
          type: string
          default: '2.0'
          example: '2.0'
        id:
          oneOf:
            - type: string
            - type: integer
          default: '1'
          example: '1'
        method:
          type: string
          enum: [tools/call]
          default: tools/call
        sessionToken:
          type: string
          description: Optional when --mcp-no-auth is active.
        params:
          type: object
          required: [name, arguments]
          properties:
            name:
              type: string
              description: MCP tool name.
            arguments:
              type: object
              description: Tool-specific arguments.
              additionalProperties: true
    JsonRpcResponse:
      type: object
      properties:
        jsonrpc:
          type: string
          example: '2.0'
        id:
          nullable: true
          oneOf:
            - type: string
            - type: integer
        result:
          nullable: true
          type: object
          properties:
            content:
              type: array
              items:
                type: object
                properties:
                  type:
                    type: string
                    example: text
                  text:
                    type: string
                    description: JSON string containing the actual tool payload.
        error:
          nullable: true
          type: object
          properties:
            code:
              type: integer
            message:
              type: string
            details:
              type: string